# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Food Order"
  description: "Restaurant food order API"
  contact:
    name: "Dev"
  license:
    name: "MIT"
host: "localhost:5000"
basePath: "/api/v1"
schemes:
    - "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /restaurants/{restaurantId}/menu:
    get:
      tags:
        - "menu"
      summary: "Retrieve restaurant menu"
      description: "Return restaurants menu which includes list of all the items that can be ordered"
      operationId: "retrieveRestaurantMenu"
      produces:
      - "application/json"
      parameters:
      - name: "restaurantId"
        in: "path"
        description: "ID of restaurant to return"
        required: true
        type: "integer"
        format: "int64"
      - name: "page"
        in: "query"
        type: "string"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/RestaurantMenu"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/ServiceException"
        "500":
          description: "Server Error"
          schema:
            $ref: "#/definitions/ServiceException"
...
definitions:
...
  RestaurantMenu:
    type: "object"
    properties:
      restaurantId:
        type: integer
      restaurants:
        type: "array"
        items:
          $ref: "#/definitions/MenuItems"
  MenuItems:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      price:
        type: "number"
      description:
        type: "string"
      photo:
        type: "string"
      status:
        type: "string"
        enum:
          - ACTIVE
          - INACTIVE
...
  ServiceException:
    type: "object"
    properties:
      code:
        type: "string"
      message:
        type: "string"
      details:
        type: "string"
...